Rename FieldClickEvent to FieldSelectEvent
Just for better description of its function
