+++ src/main/java/ships/view/BattleshipMapClickObserver.java
@@ -1,9 +1,11 @@
 package ships.view;
 
 public interface BattleshipMapClickObserver {
+
     /**
      * Method called when BattleshipMap clicked
      * @param fce
      */
-    public void fieldClickedEvent(FieldClickEvent fce, BattleshipsMap bm);
+    public void fieldClickedEvent(FieldSelectEvent fce, BattleshipsMap bm);
+
 }
